Refactor out EcoinventRelease
